+++ src/main/java/nikdevs/onlinestore/service/formatter/SizeReprFormatter.java
@@ -1,7 +1,7 @@
 package nikdevs.onlinestore.service.formatter;
 
 import nikdevs.onlinestore.service.interfaces.SizeService;
-import nikdevs.onlinestore.service.model.SizeRepr;
+import nikdevs.onlinestore.service.repr.SizeRepr;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.Formatter;
 import org.springframework.stereotype.Component;
@@ -21,11 +21,11 @@ public class SizeReprFormatter implements Formatter<SizeRepr> {
 
     @Override
     public SizeRepr parse(String s, Locale locale) throws ParseException {
-        return sizeService.findById(Integer.parseInt(s));
+        return sizeService.findById(Long.parseLong(s));
     }
 
     @Override
     public String print(SizeRepr sizeRepr, Locale locale) {
-        return Integer.toString(sizeRepr.getId());
+        return Long.toString(sizeRepr.getId());
     }
 }
