+++ src/main/java/nikdevs/onlinestore/service/interfaces/RoleService.java
@@ -1,6 +1,6 @@
 package nikdevs.onlinestore.service.interfaces;
 
-import nikdevs.onlinestore.service.model.RoleRepr;
+import nikdevs.onlinestore.service.repr.RoleRepr;
 
 import java.util.List;
 
@@ -8,8 +8,8 @@ public interface RoleService {
 
     List<RoleRepr> findAll();
     List<RoleRepr> findAllWithOutSU();
-    RoleRepr findById(int id);
+    RoleRepr findById(Long id);
     RoleRepr findByName(String name);
     void save(RoleRepr role);
-    void remove(int id);
+    void remove(Long id);
 }
